Add Generate Parameter Code Fix
