+++ src/Features/Core/Portable/FeaturesResources.resx
@@ -1623,4 +1623,10 @@ This version used in: {2}</value>
   <data name="MethodMustReturnStreamThatSupportsReadAndSeek" xml:space="preserve">
     <value>{0} must return a stream that supports read and seek operations.</value>
   </data>
+  <data name="Generate_parameter_0" xml:space="preserve">
+    <value>Generate parameter '{0}'</value>
+  </data>
+  <data name="Generate_parameter_0_and_overrides_implementations" xml:space="preserve">
+    <value>Generate parameter '{0}' (include overrides/implementations)</value>
+  </data>
 </root>
