+++ src/Features/Core/Portable/GenerateMember/GenerateVariable/AbstractGenerateVariableService.State.cs
@@ -21,6 +21,7 @@ namespace Microsoft.CodeAnalysis.GenerateMember.GenerateVariable
         {
             public INamedTypeSymbol ContainingType { get; private set; }
             public INamedTypeSymbol TypeToGenerateIn { get; private set; }
+            public IMethodSymbol ContainingMethod { get; private set; }
             public bool IsStatic { get; private set; }
             public bool IsConstant { get; private set; }
             public bool IsIndexer { get; private set; }
@@ -134,6 +135,11 @@ namespace Microsoft.CodeAnalysis.GenerateMember.GenerateVariable
             internal bool CanGenerateLocal()
             {
                 return !this.IsInMemberContext && this.IsInExecutableBlock;
+            }
+
+            internal bool CanGenerateParameter()
+            {
+                return !this.IsInMemberContext && this.ContainingMethod != null && !this.IsConstant;
             }
 
             private bool TryInitializeExplicitInterface(
@@ -260,6 +266,8 @@ namespace Microsoft.CodeAnalysis.GenerateMember.GenerateVariable
                 this.IsInConstructor = DetermineIsInConstructor(semanticDocument);
                 this.IsInMemberContext = this.SimpleNameOpt != this.SimpleNameOrMemberAccessExpressionOpt ||
                                          syntaxFacts.IsObjectInitializerNamedAssignmentIdentifier(this.SimpleNameOrMemberAccessExpressionOpt);
+
+                this.ContainingMethod = semanticModel.GetEnclosingSymbol<IMethodSymbol>(this.IdentifierToken.SpanStart, cancellationToken);
 
                 CheckSurroundingContext(semanticDocument, SymbolKind.Field, cancellationToken);
                 CheckSurroundingContext(semanticDocument, SymbolKind.Property, cancellationToken);
