+++ src/Features/Core/Portable/GenerateMember/GenerateVariable/AbstractGenerateVariableService.cs
@@ -4,6 +4,7 @@ using System.Collections.Immutable;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.AddParameter;
 using Microsoft.CodeAnalysis.CodeActions;
 using Microsoft.CodeAnalysis.CodeGeneration;
 using Microsoft.CodeAnalysis.Internal.Log;
@@ -49,29 +50,25 @@ namespace Microsoft.CodeAnalysis.GenerateMember.GenerateVariable
 
                 var canGenerateMember = CodeGenerator.CanAdd(document.Project.Solution, state.TypeToGenerateIn, cancellationToken);
 
-                // prefer fields over properties (and vice versa) depending on the casing of the member.
-                // lowercase -> fields.  title case -> properties.
-                var name = state.IdentifierToken.ValueText;
-                if (char.IsUpper(name.ToCharArray().FirstOrDefault()))
+                if (canGenerateMember)
                 {
-                    if (canGenerateMember)
+                    // prefer fields over properties (and vice versa) depending on the casing of the member.
+                    // lowercase -> fields.  title case -> properties.
+                    var name = state.IdentifierToken.ValueText;
+                    if (char.IsUpper(name.ToCharArray().FirstOrDefault()))
                     {
                         AddPropertyCodeActions(actions, semanticDocument, state);
                         AddFieldCodeActions(actions, semanticDocument, state);
                     }
-
-                    AddLocalCodeActions(actions, document, state);
-                }
-                else
-                {
-                    if (canGenerateMember)
+                    else
                     {
+
                         AddFieldCodeActions(actions, semanticDocument, state);
                         AddPropertyCodeActions(actions, semanticDocument, state);
                     }
-
-                    AddLocalCodeActions(actions, document, state);
                 }
+                AddLocalCodeActions(actions, document, state);
+                AddParameterCodeActions(actions, document, state);
 
                 if (actions.Count > 1)
                 {
@@ -177,6 +174,16 @@ namespace Microsoft.CodeAnalysis.GenerateMember.GenerateVariable
             }
         }
 
+        private void AddParameterCodeActions(ArrayBuilder<CodeAction> result, Document document, State state)
+        {
+            if (state.CanGenerateParameter())
+            {
+                result.Add(new GenerateParameterCodeAction(document, state, false));
+
+                if (AddParameterService.Instance.HasCascadingDeclarations(state.ContainingMethod))
+                    result.Add(new GenerateParameterCodeAction(document, state, true));
+            }
+        }
         private RefKind GetRefKindFromContext(State state)
         {
             if (state.IsInRefContext)
